Loop transformation refactoring: multiple loop transformation functions are moved into utils
